@@ -3,19 +3,25 @@ package cmd
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 	"io/ioutil"
 	"os"
+	"path/filepath"
+
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
 type ConfigFile struct {
-	ApiKey         string   `json:"CRONITOR_API_KEY"`
-	PingApiAuthKey string   `json:"CRONITOR_PING_API_KEY"`
-	ExcludeText    []string `json:"CRONITOR_EXCLUDE_TEXT,omitempty"`
-	Hostname       string   `json:"CRONITOR_HOSTNAME"`
-	Log            string   `json:"CRONITOR_LOG"`
-	Env            string   `json:"CRONITOR_ENV"`
+	ApiKey             string   `json:"CRONITOR_API_KEY"`
+	PingApiAuthKey     string   `json:"CRONITOR_PING_API_KEY"`
+	ExcludeText        []string `json:"CRONITOR_EXCLUDE_TEXT,omitempty"`
+	Hostname           string   `json:"CRONITOR_HOSTNAME"`
+	Log                string   `json:"CRONITOR_LOG"`
+	Env                string   `json:"CRONITOR_ENV"`
+	DashUsername       string   `json:"CRONITOR_DASH_USER"`
+	DashPassword       string   `json:"CRONITOR_DASH_PASS"`
+	AllowedIPs         string   `json:"CRONITOR_ALLOWED_IPS"`
+	CorsAllowedOrigins string   `json:"CRONITOR_CORS_ALLOWED_ORIGINS"`
 }
 
 // configureCmd represents the configure command
@@ -55,6 +61,10 @@ Example setting common exclude text for use with 'cronitor discover':
 		configData.Hostname = viper.GetString(varHostname)
 		configData.Log = viper.GetString(varLog)
 		configData.Env = viper.GetString(varEnv)
+		configData.DashUsername = viper.GetString(varDashUsername)
+		configData.DashPassword = viper.GetString(varDashPassword)
+		configData.AllowedIPs = viper.GetString(varAllowedIPs)
+		configData.CorsAllowedOrigins = viper.GetString("CRONITOR_CORS_ALLOWED_ORIGINS")
 
 		fmt.Println("\nConfiguration File:")
 		fmt.Println(configFilePath())
@@ -96,6 +106,27 @@ Example setting common exclude text for use with 'cronitor discover':
 			fmt.Println(viper.GetString(varLog))
 		}
 
+		fmt.Println("\nLocalDash Username:")
+		if configData.DashUsername == "" {
+			fmt.Println("Not Set")
+		} else {
+			fmt.Println(configData.DashUsername)
+		}
+
+		fmt.Println("\nLocalDash Password:")
+		if configData.DashPassword == "" {
+			fmt.Println("Not Set")
+		} else {
+			fmt.Println("********")
+		}
+
+		fmt.Println("\nAllowed IP Addresses:")
+		if configData.AllowedIPs == "" {
+			fmt.Println("All IPs allowed (no restrictions)")
+		} else {
+			fmt.Println(configData.AllowedIPs)
+		}
+
 		if verbose {
 			fmt.Println("\nEnviornment Variables:")
 			for _, pair := range os.Environ() {
@@ -109,27 +140,42 @@ Example setting common exclude text for use with 'cronitor discover':
 			os.Exit(1)
 		}
 
-		os.MkdirAll(defaultConfigFileDirectory(), os.ModePerm)
-		if ioutil.WriteFile(configFilePath(), b, 0644) != nil {
+		configPath := configFilePath()
+		configDir := filepath.Dir(configPath)
+
+		// Create the directory for the config file
+		if err := os.MkdirAll(configDir, os.ModePerm); err != nil {
+			fmt.Fprintf(os.Stderr,
+				"\nERROR: The configuration directory %s could not be created; check permissions and try again.\n\n",
+				configDir)
+			os.Exit(126)
+		}
+
+		if err := ioutil.WriteFile(configPath, b, 0644); err != nil {
 			fmt.Fprintf(os.Stderr,
 				"\nERROR: The configuration file %s could not be written; check permissions and try again. "+
-					   "\n       By default, configuration files are system-wide for ease of use in cron jobs and scripts. Specify an alternate config file using the --config argument or CRONITOR_CONFIG environment variable.\n\n", configFilePath())
+					"\n       By default, configuration files are system-wide for ease of use in cron jobs and scripts. Specify an alternate config file using the --config argument or CRONITOR_CONFIG environment variable.\n\n",
+				configPath)
 			os.Exit(126)
 		}
 	},
 }
 
-func configFilePath() string {
-	viperConfig := viper.ConfigFileUsed()
-	if len(viperConfig) > 0 {
-		return viperConfig
-	}
-
-	return fmt.Sprintf("%s/cronitor.json", defaultConfigFileDirectory())
-}
-
 func init() {
 	RootCmd.AddCommand(configureCmd)
 	configureCmd.Flags().StringSliceP("exclude-from-name", "e", []string{}, "Substring to always exclude from generated monitor name e.g. $ cronitor configure -e '> /dev/null' -e '/path/to/app'")
+	configureCmd.Flags().String("dash-username", "", "Username for the dashboard authentication")
+	configureCmd.Flags().String("dash-password", "", "Password for the dashboard authentication")
+	configureCmd.Flags().String("allowed-ips", "", "Comma-separated list of allowed IP addresses/CIDR ranges (e.g. 192.168.1.0/24,10.0.0.1)")
+	configureCmd.Flags().String("ping-api-key", "", "Your Cronitor Ping API key")
+	configureCmd.Flags().String("log", "", "Path to debug log file")
+	configureCmd.Flags().String("env", "", "Environment name (e.g. staging, production)")
+
 	viper.BindPFlag(varExcludeText, configureCmd.Flags().Lookup("exclude-from-name"))
+	viper.BindPFlag(varDashUsername, configureCmd.Flags().Lookup("dash-username"))
+	viper.BindPFlag(varDashPassword, configureCmd.Flags().Lookup("dash-password"))
+	viper.BindPFlag(varAllowedIPs, configureCmd.Flags().Lookup("allowed-ips"))
+	viper.BindPFlag(varPingApiKey, configureCmd.Flags().Lookup("ping-api-key"))
+	viper.BindPFlag(varLog, configureCmd.Flags().Lookup("log"))
+	viper.BindPFlag(varEnv, configureCmd.Flags().Lookup("env"))
 }
