@@ -4,45 +4,116 @@ import (
 	"bytes"
 	"compress/gzip"
 	"encoding/json"
-	"github.com/pkg/errors"
 	"fmt"
-	"github.com/getsentry/raven-go"
-	"github.com/spf13/viper"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/getsentry/raven-go"
+	"github.com/pkg/errors"
+	"github.com/spf13/viper"
 )
 
 type RuleValue string
 
 type Rule struct {
-	RuleType     string `json:"rule_type"`
+	RuleType     string    `json:"rule_type"`
 	Value        RuleValue `json:"value"`
-	TimeUnit     string `json:"time_unit,omitempty"`
-	GraceSeconds uint   `json:"grace_seconds,omitempty"`
+	TimeUnit     string    `json:"time_unit,omitempty"`
+	GraceSeconds uint      `json:"grace_seconds,omitempty"`
 }
 
+type Platform string
+
+const (
+	CRON       Platform = "cron"
+	WINDOWS    Platform = "windows"
+	KUBERNETES Platform = "kubernetes"
+	JVM        Platform = "jvm"
+	LARAVEL    Platform = "laravel"
+	MAGENTO    Platform = "magento"
+	SIDEKIQ    Platform = "sidekiq"
+	CELERY     Platform = "celery"
+	JENKINS    Platform = "jenkins"
+	QUARTZ     Platform = "quartz"
+	SPRING     Platform = "spring"
+	CLOUDWATCH Platform = "cloudwatch"
+	NODECRON   Platform = "node-cron"
+)
+
 type Monitor struct {
-	Name             string              `json:"name,omitempty"`
-	DefaultName      string              `json:"defaultName"`
-	Key              string              `json:"key"`
-	Rules            []Rule              `json:"rules"`
-	Tags             []string            `json:"tags"`
-	Type             string              `json:"type"`
-	Code             string              `json:"code,omitempty"`
-	Timezone         string              `json:"timezone,omitempty"`
-	Note             string              `json:"defaultNote,omitempty"`
-	Notifications    map[string][]string `json:"notifications,omitempty"`
-	NoStdoutPassthru bool                `json:"-"`
+	Attributes struct {
+		GroupName string `json:"group_name"`
+		Key       string `json:"key"`
+		Code      string `json:"code"`
+	} `json:"attributes,omitempty"`
+	Name             string   `json:"name,omitempty"`
+	DefaultName      string   `json:"defaultName"`
+	Key              string   `json:"key"`
+	Schedule         string   `json:"schedule,omitempty"`
+	Platform         Platform `json:"platform,omitempty"`
+	Tags             []string `json:"tags"`
+	Type             string   `json:"type"`
+	Code             string   `json:"code,omitempty"`
+	Timezone         string   `json:"timezone,omitempty"`
+	Note             string   `json:"defaultNote,omitempty"`
+	Notify           []string `json:"notify,omitempty"`
+	Passing          bool     `json:"passing,omitempty"`
+	Initialized      bool     `json:"initialized,omitempty"`
+	Disabled         bool     `json:"disabled,omitempty"`
+	Paused           *bool    `json:"paused,omitempty"`
+	NoStdoutPassthru bool     `json:"-"`
+}
+
+// UnmarshalJSON implements custom unmarshaling for the Monitor struct
+func (m *Monitor) UnmarshalJSON(data []byte) error {
+	// Create an auxiliary struct to handle the raw notify field
+	type AuxMonitor Monitor
+	aux := &struct {
+		*AuxMonitor
+		Notify interface{} `json:"notify,omitempty"`
+	}{
+		AuxMonitor: (*AuxMonitor)(m),
+	}
+
+	if err := json.Unmarshal(data, &aux); err != nil {
+		return err
+	}
+
+	// Handle the notify field
+	if aux.Notify != nil {
+		switch v := aux.Notify.(type) {
+		case []interface{}:
+			// Convert []interface{} to []string
+			m.Notify = make([]string, len(v))
+			for i, item := range v {
+				if str, ok := item.(string); ok {
+					m.Notify[i] = str
+				}
+			}
+		case []string:
+			m.Notify = v
+		case string:
+			m.Notify = []string{v}
+		}
+	}
+
+	return nil
 }
 
 type MonitorSummary struct {
 	Name        string `json:"name,omitempty"`
 	DefaultName string `json:"defaultName"`
 	Key         string `json:"key"`
-	Code        string `json:"code,omitempty"`
+	Code        string `json:"attributes.code,omitempty"`
+	Attributes  struct {
+		GroupName string `json:"group_name"`
+		Key       string `json:"key"`
+		Code      string `json:"code"`
+	} `json:"attributes,omitempty"`
 }
 
 type CronitorApi struct {
@@ -53,6 +124,10 @@ type CronitorApi struct {
 	Logger         func(string)
 }
 
+type SignupResponse struct {
+	ApiKey     string `json:"api_key"`
+	PingApiKey string `json:"ping_api_key"`
+}
 
 func (fi *RuleValue) UnmarshalJSON(b []byte) error {
 	if b[0] == '"' {
@@ -88,7 +163,7 @@ func (api CronitorApi) PutMonitors(monitors map[string]*Monitor) (map[string]*Mo
 	api.Logger("\nRequest:")
 	api.Logger(buf.String() + "\n")
 
-	response, err := api.sendHttpPut(url, jsonString)
+	response, err, _ := api.send("PUT", url, jsonString)
 	if err != nil {
 		return nil, errors.New(fmt.Sprintf("Request to %s failed: %s", url, err))
 	}
@@ -104,37 +179,55 @@ func (api CronitorApi) PutMonitors(monitors map[string]*Monitor) (map[string]*Mo
 	}
 
 	for _, value := range responseMonitors {
-		// We only need to update the Monitor struct with a code if this is a new monitor.
-		// For updates the monitor code is sent as well as the key and that takes precedence.
-		if _, ok := monitors[value.Key]; ok {
-			monitors[value.Key].Code = value.Code
+		if _, ok := monitors[value.Attributes.Key]; ok {
+			monitors[value.Attributes.Key].Attributes = value.Attributes
 		}
-
 	}
 
 	return monitors, nil
 }
 
-func (api CronitorApi) GetMonitors() ([]MonitorSummary, error) {
+func (api CronitorApi) GetMonitors() ([]Monitor, error) {
 	url := api.Url()
 	page := 1
-	monitors := []MonitorSummary{}
+	monitors := []Monitor{}
 
 	for {
-		response, err := api.sendHttpGet(fmt.Sprintf("%s?page=%d", url, page))
+		response, err, _ := api.send("GET", fmt.Sprintf("%s?page=%d", url, page), "")
+
 		if err != nil {
 			return nil, errors.New(fmt.Sprintf("Request to %s failed: %s", url, err))
 		}
 
 		type ExpectedResponse struct {
-			TotalMonitorCount int              `json:"total_monitor_count"`
-			PageSize          int              `json:"page_size"`
-			Monitors          []MonitorSummary `json:"monitors"`
+			TotalMonitorCount int       `json:"total_monitor_count"`
+			PageSize          int       `json:"page_size"`
+			Monitors          []Monitor `json:"monitors"`
 		}
 
-		responseMonitors := ExpectedResponse{}
-		if err = json.Unmarshal(response, &responseMonitors); err != nil {
+		// Create an auxiliary struct to handle raw monitor data
+		type AuxExpectedResponse struct {
+			TotalMonitorCount int               `json:"total_monitor_count"`
+			PageSize          int               `json:"page_size"`
+			Monitors          []json.RawMessage `json:"monitors"`
+		}
+
+		var auxResponse AuxExpectedResponse
+		if err = json.Unmarshal(response, &auxResponse); err != nil {
 			return nil, errors.New(fmt.Sprintf("Error from %s: %s", url, err.Error()))
+		}
+
+		// Manually unmarshal each monitor to use the custom UnmarshalJSON implementation
+		responseMonitors := ExpectedResponse{
+			TotalMonitorCount: auxResponse.TotalMonitorCount,
+			PageSize:          auxResponse.PageSize,
+			Monitors:          make([]Monitor, len(auxResponse.Monitors)),
+		}
+
+		for i, rawMonitor := range auxResponse.Monitors {
+			if err = json.Unmarshal(rawMonitor, &responseMonitors.Monitors[i]); err != nil {
+				return nil, errors.New(fmt.Sprintf("Error unmarshaling monitor: %s", err.Error()))
+			}
 		}
 
 		monitors = append(monitors, responseMonitors.Monitors...)
@@ -149,83 +242,54 @@ func (api CronitorApi) GetMonitors() ([]MonitorSummary, error) {
 }
 
 func (api CronitorApi) GetRawResponse(url string) ([]byte, error) {
-	client := &http.Client{}
-	request, err := http.NewRequest("GET", url, nil)
-	request.SetBasicAuth(viper.GetString(api.ApiKey), "")
-	request.Header.Add("Content-Type", "application/json")
-	request.Header.Add("User-Agent", api.UserAgent)
-	response, err := client.Do(request)
-	if err != nil {
-		return nil, err
-	}
-
-	if response.StatusCode != 200 {
-		return nil, errors.New(fmt.Sprintf("Unexpected %d API response", response.StatusCode))
-	}
-
-	defer response.Body.Close()
-	contents, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		raven.CaptureErrorAndWait(err, nil)
-		return nil, err
-	}
-
-	return contents, nil
+	response, err, _ := api.send("GET", url, "")
+	return response, err
 }
 
 func (api CronitorApi) Url() string {
 	if api.IsDev {
-		return "http://dev.cronitor.io/v3/monitors"
+		return "http://dev.cronitor.io/api/monitors"
 	} else {
-		return "https://cronitor.io/v3/monitors"
+		return "https://cronitor.io/api/monitors"
 	}
 }
 
-func (api CronitorApi) sendHttpPut(url string, body string) ([]byte, error) {
+func (api CronitorApi) send(method string, url string, body string) ([]byte, error, int) {
 	client := &http.Client{
 		Timeout: 120 * time.Second,
 	}
-	request, err := http.NewRequest("PUT", url, strings.NewReader(body))
-	request.SetBasicAuth(viper.GetString(api.ApiKey), "")
-	request.Header.Add("Content-Type", "application/json")
+	request, err := http.NewRequest(method, url, strings.NewReader(body))
+
+	// Always fetch the latest API key from viper to pick up settings changes
+	currentApiKey := viper.GetString("CRONITOR_API_KEY")
+	if currentApiKey == "" {
+		// Fallback to the API key stored in the struct if viper doesn't have it
+		currentApiKey = api.ApiKey
+	}
+	request.SetBasicAuth(currentApiKey, "")
+
+	if strings.HasSuffix(url, "/signup") || strings.HasSuffix(url, "/sign-up") {
+		request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
+	} else {
+		request.Header.Add("Content-Type", "application/json")
+	}
+
 	request.Header.Add("User-Agent", api.UserAgent)
+	request.Header.Add("Cronitor-Version", "2020-10-01")
 	request.ContentLength = int64(len(body))
 	response, err := client.Do(request)
 	if err != nil {
-		return nil, err
+		return nil, err, 0
 	}
 
 	defer response.Body.Close()
 	contents, err := ioutil.ReadAll(response.Body)
 	if err != nil {
 		raven.CaptureErrorAndWait(err, nil)
-		return nil, err
+		return nil, err, 0
 	}
 
-	return contents, nil
-}
-
-func (api CronitorApi) sendHttpGet(url string) ([]byte, error) {
-	client := &http.Client{
-		Timeout: 120 * time.Second,
-	}
-	request, err := http.NewRequest("GET", url, nil)
-	request.SetBasicAuth(viper.GetString(api.ApiKey), "")
-	request.Header.Add("Content-Type", "application/json")
-	request.Header.Add("User-Agent", api.UserAgent)
-	response, err := client.Do(request)
-	if err != nil {
-		return nil, err
-	}
-
-	defer response.Body.Close()
-	contents, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		raven.CaptureErrorAndWait(err, nil)
-		return nil, err
-	}
-
-	return contents, nil
+	return contents, nil, response.StatusCode
 }
 
 func gzipLogData(logData string) *bytes.Buffer {
@@ -249,28 +313,17 @@ func gzipLogData(logData string) *bytes.Buffer {
 func getPresignedUrl(apiKey string, postBody []byte) ([]byte, error) {
 	url := "https://cronitor.io/api/logs/presign"
 
-	client := &http.Client{Timeout: 120 * time.Second}
-	request, err := http.NewRequest("POST", url, strings.NewReader(string(postBody)))
-	if err != nil {
-		return nil, errors.Wrap(err, "could not create request for URL presign")
+	api := CronitorApi{
+		ApiKey:    apiKey,
+		UserAgent: "cronitor-cli",
 	}
-	request.SetBasicAuth(apiKey, "")
-	request.Header.Add("Content-Type", "application/json")
-	response, err := client.Do(request)
+
+	response, err, _ := api.send("POST", url, string(postBody))
 	if err != nil {
 		return nil, errors.Wrap(err, "error requesting presigned url")
 	}
-	if response.StatusCode != 200 && response.StatusCode != 201 {
-		return nil, fmt.Errorf("error response code %d returned", response.StatusCode)
-	}
 
-	contents, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		return nil, err
-	}
-	defer response.Body.Close()
-	response.Body = ioutil.NopCloser(bytes.NewBuffer(contents))
-	return contents, nil
+	return response, nil
 }
 
 func SendLogData(apiKey string, monitorKey string, seriesID string, outputLogs string) ([]byte, error) {
@@ -282,6 +335,7 @@ func SendLogData(apiKey string, monitorKey string, seriesID string, outputLogs s
 	if err != nil {
 		return nil, errors.Wrap(err, "couldn't encode job and series IDs to JSON")
 	}
+
 	var responseJson struct {
 		Url string `json:"url"`
 	}
@@ -292,6 +346,7 @@ func SendLogData(apiKey string, monitorKey string, seriesID string, outputLogs s
 	if err := json.Unmarshal(response, &responseJson); err != nil {
 		return nil, err
 	}
+
 	s3LogPutUrl := responseJson.Url
 	if len(s3LogPutUrl) == 0 {
 		return nil, errors.New("no presigned S3 url returned. Something is wrong")
@@ -303,18 +358,76 @@ func SendLogData(apiKey string, monitorKey string, seriesID string, outputLogs s
 	client := &http.Client{
 		Timeout: 120 * time.Second,
 	}
-	response2, err := client.Do(req)
-	if err != nil || response == nil {
-		return nil, errors.Wrap(err, fmt.Sprintf("error putting logs: %v", response2))
-	}
-	if response2.StatusCode < 200 || response2.StatusCode >= 300 {
-		return nil, fmt.Errorf("error response code %d returned", response2.StatusCode)
-	}
-	body, err := ioutil.ReadAll(response2.Body)
+	resp, err := client.Do(req)
 	if err != nil {
 		return nil, err
 	}
-	defer response2.Body.Close()
-	//log(fmt.Sprintf("logs shipped for series %s", seriesID))
-	return body, nil
+	defer resp.Body.Close()
+
+	return ioutil.ReadAll(resp.Body)
+}
+
+func (api CronitorApi) Signup(name string, email string, password string) (*SignupResponse, error) {
+	payload := fmt.Sprintf("fullname=%s&email=%s&password=%s",
+		url.QueryEscape(name),
+		url.QueryEscape(email),
+		url.QueryEscape(password))
+
+	url := "https://cronitor.io/sign-up"
+	if api.IsDev {
+		url = "http://dev.cronitor.io/sign-up"
+	}
+
+	response, err, statusCode := api.send("POST", url, payload)
+	if err != nil {
+		return nil, err
+	}
+
+	if statusCode != 200 {
+		return nil, fmt.Errorf("sign up failed (status %d): %s", statusCode, string(response))
+	}
+
+	if statusCode != 200 {
+		return nil, fmt.Errorf("sign up failed: %d", statusCode)
+	}
+
+	var signupResp SignupResponse
+	if err := json.Unmarshal(response, &signupResp); err != nil {
+		return nil, fmt.Errorf("failed to parse signup response: %s", err)
+	}
+
+	return &signupResp, nil
+}
+
+// PauseMonitor pauses a monitor for a specified number of hours.
+// If hours is empty, the monitor is paused permanently.
+func (c *CronitorApi) PauseMonitor(code string, hours string) error {
+	url := fmt.Sprintf("%s/%s/pause", c.Url(), code)
+	if hours != "" {
+		url = fmt.Sprintf("%s/%s", url, hours)
+	}
+	fmt.Println("url", url)
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return err
+	}
+
+	// Always fetch the latest API key from viper to pick up settings changes
+	currentApiKey := viper.GetString("CRONITOR_API_KEY")
+	if currentApiKey == "" {
+		// Fallback to the API key stored in the struct if viper doesn't have it
+		currentApiKey = c.ApiKey
+	}
+	req.SetBasicAuth(currentApiKey, "")
+
+	_, err, statusCode := c.send("GET", url, "")
+	if err != nil {
+		return err
+	}
+
+	if statusCode != http.StatusOK {
+		return fmt.Errorf("failed to pause monitor: status code %d", statusCode)
+	}
+
+	return nil
 }
